chore: move DisplayConditionSettings to legacy
